+++ CHANGES.md
@@ -12,6 +12,7 @@
 * Lib: make the Wasm version of Json.output work with native ints and JavaScript objects (#1872)
 * Runtime: implement BLAKE2b primitives for Wasm (#1873)
 * Runtime: delete BigStringReader, one should use UInt8ArrayReader instead
+* Runtime/wasm: support jsoo_env and keep track of backtrace status (#1881)
 
 ## Bug fixes
 * Runtime: fix path normalization (#1848)